+++ MonacoEditorComponent/Monaco/Uri.cs
@@ -6,7 +6,8 @@ namespace Monaco
     /// Uniform Resource Identifier (Uri) http://tools.ietf.org/html/rfc3986.
     /// This class is a simple parser which creates the basic component parts
     /// (http://tools.ietf.org/html/rfc3986#section-3) with minimal validation
-    /// and encoding.///       foo://example.com:8042/over/there?name=ferret#nose
+    /// and encoding.
+    ///       foo://example.com:8042/over/there?name=ferret#nose
     ///       \_/   \______________/\_________/ \_________/ \__/
     ///        |           |            |            |        |
     ///     scheme     authority       path        query   fragment